requirements.txt
SQLAlchemy==1.4.49
streamlit==1.38.0
streamlit-authenticator==0.3.3
PyYAML==6.0.1